Update StreamConnection.cs
Up to 10% of received activities were being lost. The main culprit behind it was line 86:

if (!output.Contains(0x0A)) continue;  //Heartbeat

Which was skipping iteration without clearing the allocated memory hence causing memory in different threads to clash and merge in part or whole.

I have also added some other parts to make it a good commercial base for those starting out on PowerTrack. I hope it will be of benefit to GNIP and its customers.

Mustafa Ozkececigil/Fluxifi